system/extras/libpagemap: Fix calculation for process' swapped pages
The current logic that interprets the pagemap value provided
by the kernel considers that a page is not part of a process'
memory if the page is not present. However, pages that are
swapped out are part of a process' memory, but they are not
present, and are thus not accounted for as part of a process'
swap usage.

Fix this by ignoring pages that are both not present and not
swapped when calculating a process' memory consumption.

Change-Id: I10d5614d075ded18a910c0fb74f65be57f572f91