simpleperf: Fix auxtrace data location check.
When trying to read the content of an aux record from the auxtrace
records, make sure that the found auxtrace record actually contains
the aux record's data, and is not just what happens to be closest to
it. Neglecting to do so can lead to reading nonsense data or even
trying to read outside the perf.data file.

A test is added to make sure this works. The test file is a copy of
perf_etm.data, with one aux record's offset modified (by hand) to
point outside the file. This causes simpleperf to exit with an
error (making the test fail), but with the change it proceeds with an
info message. While the test data was manually crafted, this situation
can "naturally" arise when auxtrace records are discarded because they
do not fit the userspace buffer. However, perf.data like those are
usually large; the hand-crafted test data was chosen for its small
size.

Test: simpleperf_unit_test --gtest_filter=cmd_dump.*
Change-Id: I036dedfa7c51ee515771ddc0b021920b6c2a4e2f