Add Bazel build system support
Allow developers to build Torq without needing to setup an entire
Android source tree. With Bazel, developers can build torq with just the
torq code source itself and no dependency to Android.

Also, refactored the README to be up-to-date with latest changes and
have a better UI.

Test: ./tools/torq_test
Flag: EXEMPT refactor
Bug: 417561323
Change-Id: I9ec0fd5d0cd330a495d336df7f94041c290d3908