profcollect: Set our own property to trigger init to start the daemon
We relied on device_config to set flag values under the
persist.device_config.aconfig_flags.profcollect_native_boot namespace to
trigger init to start the profcollectd daemon. We have observed that
this flag is often not properly set to reflect the server provisioned
flag value.

This change adds a command to profcollectctl to set our own system
property. The property is synchronized with the server flag during
system boot. profcollectd starts as soon as the property becomes true,
and resets itself once it becomes false.

Test: N/A
Bug: 423703735
Change-Id: I52a91255d334d55201da8794610f36a0e4ebfbb4