Prevent torq from stalling when no file server request is received
Torq opens a socketserver to serve the trace file to the browser to be
loaded by the Perfetto UI. However, if the tab is closed before a
request is made, torq will stall and cannot be closed without shutting
down the terminal. Add a timeout to prevent this stall.

Test: ./torq_test
Flag: EXEMPT refactor
Fixes: 427999031
Change-Id: Id1e948e79d077ad8b0cbf2c7ef538741763e77e3