Update torq's simpleperf implementation to work with Bazel
Bazel is configured to work in a sandbox environment. Thus, when
subprocesses are started in Bazel, the directory that the scripts are
being run from are not added to sys.path because of the sandbox
environment. This means that the simpleperf scripts that are downloaded
for torq are unable to import files from their own directory because
that directory is not a part of sys.path. Load this directory into the
PYTHONPATH variable before calling the scripts to fix this error.

Test: ./torq_test
Flag: EXEMPT refactor
Bug: 433823065
Change-Id: I98d916d5323a9097f175f01411be1b9b82f85a6f