Add a subprocess.run wrapper to throw when an error is found in torq
In torq, when subprocess.run is called, it will ignore errors given by
the commands it's running. This can lead to missing errors in the code.
Add a wrapper function that will throw on non-zero return codes and can
take a list of return codes to ignore for edge cases where a command
error shouldn't result in throwing. Add a generate_mock_process function
to test_utils to handle the returncodes passed to run_subprocess.

Test: ./torq_test
Flag: EXEMPT refactor
Fixes: 439910552
Change-Id: I337644e1f065386e14d56fc52b04e48fc2cfbaf1