boottime: Add support for measuring launcher shown time
This change introduces the capability to measure the time it takes for
the launcher to be shown. This functionality is crucial for more
accurate boot time measurement.

A new command-line flag, `--wait-launcher-shown`, has been added to
`bootanalyze.py`. When this flag is used, the script will wait for the
"LAUNCHER_SHOWN" logcat event and record its timing.

A new metric, `*LauncherShown+Bootloader`, is also introduced to provide
the more accurate measurement for time from power_on to login.

Bug: 417581104
Test: ./bootanalyzer.py --wait-launcher-shown works LauncherShown event
recorded

Change-Id: I4638657f643925ea646e9f45126c68985f1f8dcf