simpleperf: stat: Fix --monitor-new-thread for multi-process apps
When using `--monitor-new-thread --per-thread --app <app_package_name>`,
the stat command may only monitor new threads for one process in
the app, while the app may have mutliple processes.

This was because the initial set of monitored processes was determined
at the start, and new processes spawned by the app were not added to
the monitored set.

This patch fixes the issue by periodically scanning for new processes
belonging to the monitored app package and adding them to the monitored
process set. This ensures that new threads from any process in the app
are detected and monitored.

Bug: 448177171
Test: run simpleperf stat command
Test: run simpleperf_unit_test
Change-Id: I6a87ec352e896b0caa92bca0f48da9addaa5d3c2