simpleperf: Fix for creating AuxBuffer in case of multiple events
If there are multiple events enabled on the same CPU, both MappedBuffer
and AuxBuffer are created only for the first event in the list, for the
rest the MappedBuffer is shared but the AuxBuffer is not.
This casues a problem if the second (or a later) event requires an
AuxBuffer.
Fix is to sort the events so that all the events that require AuxBuffer
come first.

Test: Manual testing, with multiple events where HW/PMU events come
before SPE, using the folowing (and similar) commands:
simpleperf record -e instructions,cpu-cycles -e arm_spe/ -- <workload>
Cherrypick-From: https://partner-android-review.git.corp.google.com/c/platform/system/extras/+/3569554

Bug: 467135836
Change-Id: I4d59fbe6c7b70a8145e16964020c0565e801f924