simpleperf: Prefer simpleperf_app_runner on Android 16+
Previously, simpleperf always preferred `simpleperf_app_runner`, which
was not optimal for older Android versions where `run-as` allows using
a newer, sideloaded simpleperf with full capabilities.

Starting with Android 16, sideloaded simpleperf binaries lack the
permissions required to collect kernel samples. To address this,
this change prioritizes `simpleperf_app_runner` on Android 16 and
later. This ensures we use the system-image version of simpleperf,
which retains the necessary permissions.

For pre-Android 16 devices, continue to prefer `run-as`, allowing
the use of updated, sideloaded simpleperf binaries, as this was the
intended and more effective approach for those versions.

Additionally, add support for detecting Android 16 (codename 'B',
SDK level 36).

Bug: 455595418
Test: atest simpleperf_unit_test
Change-Id: I002862fde34997b79155a11f73e73f5719a03c9c