simpleperf: app_profiler.py: Check if app is debug build
Profiling debug builds often yields performance data that is not
representative of the release version. To prevent accidental profiling
of debug builds, this change updates app_profiler.py to check the
application type on Android >= 12.

If a debug build is detected, the profiler will now abort with an error.
Users can override this behavior and force profiling of a debug build
by using the new flag: --unrepresentative_profile_debug_app.

Additionally:
- Updated documentation to recommend profiling release builds configured
  with <profileable android:shell="true" />.
- Updated tests to use the new flag where appropriate.
- Added a new test case in TestRecordingRealApps to verify that
  profiling a debug build without the flag fails as expected.

Bug: 469208402
Test: run test.py
Change-Id: Ie35e55ffc91812dc3d7201bcadd787e1f433e59c